pyproject.toml
[project]
name = "langgraph-websocket-test"
version = "0.1.0"
description = "AI-powered website screenshot analyzer using LangGraph, Meta Llama Vision, and Steel.dev"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=1.0.1",
    "langchain>=1.0.2",
    "langchain-openai>=1.0.1",
    "fastapi>=0.119.1",
    "uvicorn[standard]>=0.38.0",
    "python-dotenv>=1.1.1",
    "openai>=2.6.0",
    "steel-sdk==0.13.0",
    "httpx>=0.28.1",
    "google-search-results>=2.4.2",
    "boto3>=1.40.58",
    "supabase>=2.22.1",
    "clerk-backend-api>=3.3.1",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "python-socketio>=5.14.2",
    "redis>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = []
only-include = ["server.py", "client.html"]
